add parseLinks to parse all links in a document
this will initially be used for finding potential candidates to send
webmentions to, as well as performing the verification step after
receiving a webmention.

in theory this could be combined with htmlLink, but they are different
enough that they are separate for now.  For example, htmlLink gives
preference to <link> elements over <a> elements, and returns as soon as
a matching link is found.  parseLinks parses the full document every
time.