simplify html parsing logic for endpoint discovery
The webmention spec explicitly states that <link> elements should be
preferred over <a> elements when discovering the webmention endpoint.
However, the HTML5 spec also makes it very clear that <link> elements
containing a rel attribute may only appear in the document's <head>,
while <a> elements may only appear in the document's <body>.  Therefore,
assuming well-formed documents, simply searching for the first <a> or
<link> element in document order with the webmention rel value will give
the desired result.